Move LSPViewEventListener, remove global_events from formatting.py
