+++ plugin/formatting.py
@@ -1,14 +1,12 @@
 import sublime
-import sublime_plugin
 from .core.protocol import Request
 from .core.configurations import is_supported_syntax
 from .core.settings import client_configs
 from .core.edit import parse_text_edit
-from .core.registry import LspTextCommand, session_for_view, client_from_session, sessions_for_view
+from .core.registry import LspTextCommand, LSPViewEventListener, session_for_view, client_from_session, sessions_for_view
 from .core.url import filename_to_uri
 from .core.sessions import Session
 from .core.views import region_to_range
-from .core.events import global_events
 
 try:
     from typing import Dict, Any, List, Optional
@@ -34,43 +32,7 @@ def wants_will_save_wait_until(session: Session) -> bool:
     return False
 
 
-def run_will_save_wait_until(view: sublime.View, file_path: str, session: Session) -> None:
-    client = client_from_session(session)
-    if client:
-        # Make sure that the server sees the most recent document changes.
-        global_events.publish("view.on_purge_changes", view)
-        params = {
-            "textDocument": {
-                "uri": filename_to_uri(file_path)
-            },
-            "reason": 1  # TextDocumentSaveReason.Manual
-        }
-        request = Request.willSaveWaitUntil(params)
-        response = client.execute_request(request)
-        if response:
-            apply_response_to_view(response, view)
-            global_events.publish("view.on_purge_changes", view)
-
-
-def run_format_on_save(view: sublime.View, file_path: str) -> None:
-    client = client_from_session(session_for_view(view, 'documentFormattingProvider'))
-    if client:
-        # Make sure that the server sees the most recent document changes.
-        global_events.publish("view.on_purge_changes", view)
-        params = {
-            "textDocument": {
-                "uri": filename_to_uri(file_path)
-            },
-            "options": options_for_view(view)
-        }
-        request = Request.formatting(params)
-        response = client.execute_request(request)
-        if response:
-            apply_response_to_view(response, view)
-            global_events.publish("view.on_purge_changes", view)
-
-
-class FormatOnSaveListener(sublime_plugin.ViewEventListener):
+class FormatOnSaveListener(LSPViewEventListener):
     def __init__(self, view: sublime.View) -> None:
         super().__init__(view)
 
@@ -86,12 +48,50 @@ class FormatOnSaveListener(sublime_plugin.ViewEventListener):
         if not file_path:
             return
 
+        self._view_maybe_dirty = True
         for session in sessions_for_view(self.view):
             if wants_will_save_wait_until(session):
-                run_will_save_wait_until(self.view, file_path, session)
+                self._purge_changes_if_needed()
+                self._will_save_wait_until(file_path, session)
 
         if self.view.settings().get("lsp_format_on_save"):
-            run_format_on_save(self.view, file_path)
+            self._purge_changes_if_needed()
+            self._format_on_save(file_path)
+
+    def _purge_changes_if_needed(self) -> None:
+        if self._view_maybe_dirty:
+            self.manager._documents.purge_changes(self.view)
+            self._view_maybe_dirty = False
+
+    def _will_save_wait_until(self, file_path: str, session: Session) -> None:
+        client = client_from_session(session)
+        if client:
+            params = {
+                "textDocument": {
+                    "uri": filename_to_uri(file_path)
+                },
+                "reason": 1  # TextDocumentSaveReason.Manual
+            }
+            request = Request.willSaveWaitUntil(params)
+            response = client.execute_request(request)
+            if response:
+                apply_response_to_view(response, self.view)
+                self._view_maybe_dirty = True
+
+    def _format_on_save(self, file_path: str) -> None:
+        client = client_from_session(session_for_view(self.view, 'documentFormattingProvider'))
+        if client:
+            params = {
+                "textDocument": {
+                    "uri": filename_to_uri(file_path)
+                },
+                "options": options_for_view(self.view)
+            }
+            request = Request.formatting(params)
+            response = client.execute_request(request)
+            if response:
+                apply_response_to_view(response, self.view)
+                self._view_maybe_dirty = True
 
 
 class LspFormatDocumentCommand(LspTextCommand):
